considerable improvement in initialization logic;  method of playing short clips (with aid=4) has seemingly been perfected.
we are still massively over-initializing.  For instance we will
bring up HEOS or the Hue bridge even when we don't use it.  Sometimes
it hangs for a second or two (bad enough) but I suspect also that having
more coroutines around means more things can go wrong.

The GUI is almost certainly broken,  but I am planning on fixing it
once I have initialization defeated in the CLI.