fix logic for fetching motd images.
with the original logic,  the system did not appear to be reliably
showing the last image.

rather than debugging the original logic,  i corrected it to better
fit the specification.  Specifically,  I was doing a HEAD request
and then later doing a GET request,  which is not necessarily
Kosher because of the possible gap in time between those two
requests.

Note that the motd logic is similar but different to "GET with cache"
as implemented for the text weather forecasts.  In the case of the
motd we are archiving all of the images we get (if only that is
because we've archived other images so far.).  Consolidating these
two use caches (also binary and text fetch) is another project.